[AT Services] Only try to remove /tmp/atBinder if it exists
The app doesn't start properly at the moment since it tries to
remove a file that doesn't exist.

Resolves: LE-6607

+++ components/atServices/atBinder/le_atBinder.c
@@ -36,10 +36,13 @@ COMPONENT_INIT
 
     LE_INFO("atBinder starts");
 
-    if (unlink(AT_BINDER_SOCK_PATH) < 0)
+    if (0 == access(AT_BINDER_SOCK_PATH, F_OK))
     {
-        LE_ERROR("unlink socket failed: %m");
-        goto exit;
+        if (unlink(AT_BINDER_SOCK_PATH) < 0)
+        {
+            LE_ERROR("unlink socket failed: %m");
+            goto exit;
+        }
     }
 
     // Create the socket
